backend/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
numpy
opencv-python
python-multipart
httpx
pandas
